Move PhaseEvent and PhaseEventLevel to a separate file
