+++ src/stores/flyer.ts
@@ -9,7 +9,8 @@ import type { Flyer } from "../data/Flyer"
 import type { FlyerSettings } from "../data/FlyerSettings"
 import type { IScheduler } from "../data/IScheduler"
 import { KnockoutScheduler } from "../data/KnockoutScheduler"
-import { type Phase, type PhaseEvent, PhaseEventLevel } from "../data/Phase"
+import type { Phase } from "../data/Phase"
+import { PhaseEventLevel, type PhaseEvent } from "../data/PhaseEvent"
 import { createPlayOffSettings, Format } from "../data/PhaseSettings"
 import type { Player } from "../data/Player"
 import type { PlayOff } from "../data/PlayOff"
